Struct cupid::Master

pub struct Master { /* fields omitted */ }

Information about the currently running processor

Feature flags match the feature mnemonic listed in the Intel Instruction Set Reference. This struct provides a facade for flags so the consumer doesn't need to worry about which particular CPUID leaf provides the information.

For data beyond simple feature flags, you will need to retrieve the nested struct and call the appropriate methods on it.

Methods

impl Master

fn version_information(&self) -> Option<&VersionInformation>

fn thermal_power_management_information(&self)
-> Option<&ThermalPowerManagementInformation>

fn structured_extended_information(&self)
-> Option<&StructuredExtendedInformation>

fn extended_processor_signature(&self) -> Option<&ExtendedProcessorSignature>

fn cache_line(&self) -> Option<&CacheLine>

fn time_stamp_counter(&self) -> Option<&TimeStampCounter>

fn physical_address_size(&self) -> Option<&PhysicalAddressSize>

fn brand_string(&self) -> Option<&str>

fn sse3(&self) -> bool

fn pclmulqdq(&self) -> bool

fn dtes64(&self) -> bool

fn monitor(&self) -> bool

fn ds_cpl(&self) -> bool

fn vmx(&self) -> bool

fn smx(&self) -> bool

fn eist(&self) -> bool

fn tm2(&self) -> bool

fn ssse3(&self) -> bool

fn cnxt_id(&self) -> bool

fn sdbg(&self) -> bool

fn fma(&self) -> bool

fn cmpxchg16b(&self) -> bool

fn xtpr_update_control(&self) -> bool

fn pdcm(&self) -> bool

fn pcid(&self) -> bool

fn dca(&self) -> bool

fn sse4_1(&self) -> bool

fn sse4_2(&self) -> bool

fn x2apic(&self) -> bool

fn movbe(&self) -> bool

fn popcnt(&self) -> bool

fn tsc_deadline(&self) -> bool

fn aesni(&self) -> bool

fn xsave(&self) -> bool

fn osxsave(&self) -> bool

fn avx(&self) -> bool

fn f16c(&self) -> bool

fn rdrand(&self) -> bool

fn fpu(&self) -> bool

fn vme(&self) -> bool

fn de(&self) -> bool

fn pse(&self) -> bool

fn tsc(&self) -> bool

fn msr(&self) -> bool

fn pae(&self) -> bool

fn mce(&self) -> bool

fn cx8(&self) -> bool

fn apic(&self) -> bool

fn sep(&self) -> bool

fn mtrr(&self) -> bool

fn pge(&self) -> bool

fn mca(&self) -> bool

fn cmov(&self) -> bool

fn pat(&self) -> bool

fn pse_36(&self) -> bool

fn psn(&self) -> bool

fn clfsh(&self) -> bool

fn ds(&self) -> bool

fn acpi(&self) -> bool

fn mmx(&self) -> bool

fn fxsr(&self) -> bool

fn sse(&self) -> bool

fn sse2(&self) -> bool

fn ss(&self) -> bool

fn htt(&self) -> bool

fn tm(&self) -> bool

fn pbe(&self) -> bool

fn digital_temperature_sensor(&self) -> bool

fn intel_turbo_boost(&self) -> bool

fn arat(&self) -> bool

fn pln(&self) -> bool

fn ecmd(&self) -> bool

fn ptm(&self) -> bool

fn hwp(&self) -> bool

fn hwp_notification(&self) -> bool

fn hwp_activity_window(&self) -> bool

fn hwp_energy_performance_preference(&self) -> bool

fn hdc(&self) -> bool

fn hardware_coordination_feedback(&self) -> bool

fn performance_energy_bias(&self) -> bool

fn fsgsbase(&self) -> bool

fn ia32_tsc_adjust_msr(&self) -> bool

fn bmi1(&self) -> bool

fn hle(&self) -> bool

fn avx2(&self) -> bool

fn smep(&self) -> bool

fn bmi2(&self) -> bool

fn enhanced_rep_movsb_stosb(&self) -> bool

fn invpcid(&self) -> bool

fn rtm(&self) -> bool

fn pqm(&self) -> bool

fn deprecates_fpu_cs_ds(&self) -> bool

fn pqe(&self) -> bool

fn rdseed(&self) -> bool

fn adx(&self) -> bool

fn smap(&self) -> bool

fn intel_processor_trace(&self) -> bool

fn prefetchwt1(&self) -> bool

fn lahf_sahf_in_64_bit(&self) -> bool

fn lzcnt(&self) -> bool

fn prefetchw(&self) -> bool

fn syscall_sysret_in_64_bit(&self) -> bool

fn execute_disable(&self) -> bool

fn gigabyte_pages(&self) -> bool

fn rdtscp_and_ia32_tsc_aux(&self) -> bool

fn intel_64_bit_architecture(&self) -> bool

fn invariant_tsc(&self) -> bool